Make presenter concept generic
