+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/runInTerminalHelpers.ts
@@ -47,44 +47,6 @@ export function isFish(envShell: string, os: OperatingSystem): boolean {
 	return /^fish$/.test(pathPosix.basename(envShell));
 }
 
-/**
- * Extracts the Python code from a `python -c "..."` or `python -c '...'` command,
- * returning the code with properly unescaped quotes.
- *
- * @param commandLine The full command line to parse
- * @param shell The shell path (to determine quote escaping style)
- * @param os The operating system
- * @returns The extracted Python code, or undefined if not a python -c command
- */
-export function extractPythonCommand(commandLine: string, shell: string, os: OperatingSystem): string | undefined {
-	// Match python/python3 -c "..." pattern (double quotes)
-	const doubleQuoteMatch = commandLine.match(/^python(?:3)?\s+-c\s+"(?<python>.+)"$/s);
-	if (doubleQuoteMatch?.groups?.python) {
-		let pythonCode = doubleQuoteMatch.groups.python.trim();
-
-		// Unescape quotes based on shell type
-		if (isPowerShell(shell, os)) {
-			// PowerShell uses backtick-quote (`") to escape quotes inside double-quoted strings
-			pythonCode = pythonCode.replace(/`"/g, '"');
-		} else {
-			// Bash/sh/zsh use backslash-quote (\")
-			pythonCode = pythonCode.replace(/\\"/g, '"');
-		}
-
-		return pythonCode;
-	}
-
-	// Match python/python3 -c '...' pattern (single quotes)
-	// Single quotes in bash/sh/zsh are literal - no escaping inside
-	// Single quotes in PowerShell are also literal
-	const singleQuoteMatch = commandLine.match(/^python(?:3)?\s+-c\s+'(?<python>.+)'$/s);
-	if (singleQuoteMatch?.groups?.python) {
-		return singleQuoteMatch.groups.python.trim();
-	}
-
-	return undefined;
-}
-
 // Maximum output length to prevent context overflow
 const MAX_OUTPUT_LENGTH = 60000; // ~60KB limit to keep context manageable
 export const TRUNCATION_MESSAGE = '\n\n[... PREVIOUS OUTPUT TRUNCATED ...]\n\n';
